refactor: move files in svelte to components
