+++ src/components/Quote.svelte
