+++ src/components/Verdict.svelte
